Update zpathstore implementation to support zip files in loadpaths, Updated getsource to support zipfiles and add to various primitives. Update getsource and loadmods to work with zip archives, add getsource support to various primitives
Updated testing to include ziptests

+++ src/sys/config.c
@@ -400,7 +400,7 @@ static lispval file_config_lookup(lispval symbol,void *pathdata)
   u8_string filename =
     u8_find_file(SYM_NAME(symbol),path,u8_file_readablep);
   if (filename) {
-    int n_bytes; lispval results = KNO_EMPTY;
+    ssize_t n_bytes; lispval results = KNO_EMPTY;
     unsigned char *content = u8_filedata(filename,&n_bytes);
     if (content[0]==0) {
       struct KNO_INBUF in = { 0 };
